Handling nulls in Records and UserDetails

+++ src/main/java/codeenthusiast/TrainingCenterApp/record/custom/CustomRecordServiceImpl.java
@@ -1,5 +1,6 @@
 package codeenthusiast.TrainingCenterApp.record.custom;
 
+import codeenthusiast.TrainingCenterApp.exceptions.EntityNotFoundException;
 import codeenthusiast.TrainingCenterApp.record.PersonalRecords;
 import codeenthusiast.TrainingCenterApp.record.PersonalRecordsServiceImpl;
 import codeenthusiast.TrainingCenterApp.security.services.UserDetailsImpl;
@@ -30,6 +31,8 @@ public class CustomRecordServiceImpl implements CustomRecordService {
     @Override
     public CustomRecordDTO createCustomRecord(Long personalRecordsId, CustomRecordDTO customRecordDTO) {
         PersonalRecords personalRecords = personalRecordsServiceImpl.getPersonalRecordsById(personalRecordsId);
+        if(isNull(personalRecords))
+            throw new EntityNotFoundException("Resource not available");
         if(!personalRecordsServiceImpl.hasAccess(personalRecords))
             throw new AccessDeniedException("Access denied");
         CustomRecord customRecord = mapToEntity(customRecordDTO);
@@ -40,6 +43,8 @@ public class CustomRecordServiceImpl implements CustomRecordService {
     @Override
     public CustomRecordDTO updateCustomRecord(Long customRecordId, CustomRecordDTO customRecordDTO) {
         CustomRecord customRecord = getCustomRecordByIdFromRepo(customRecordId);
+        if(isNull(customRecord))
+            throw new EntityNotFoundException("Resource not available");
         if(!hasAccess(customRecord))
             throw new AccessDeniedException("Access denied");
         updateCustomRecord(customRecord, customRecordDTO);
@@ -49,6 +54,8 @@ public class CustomRecordServiceImpl implements CustomRecordService {
     @Override
     public List<CustomRecordDTO> getAllCustomRecordsByPersonalRecordsId(Long personalRecordsId) {
         PersonalRecords personalRecords = personalRecordsServiceImpl.getPersonalRecordsById(personalRecordsId);
+        if(isNull(personalRecords))
+            throw new EntityNotFoundException("Resource not available");
         if(!personalRecordsServiceImpl.hasAccess(personalRecords))
             throw new AccessDeniedException("Access denied");
         return mapToDTOs(repository.findAllByPersonalRecordsId(personalRecordsId));
@@ -57,6 +64,8 @@ public class CustomRecordServiceImpl implements CustomRecordService {
     @Override
     public List<CustomRecordDTO> getThreeLatestCustomRecordsByPersonalRecordsId(Long personalRecordsId) {
         PersonalRecords personalRecords = personalRecordsServiceImpl.getPersonalRecordsById(personalRecordsId);
+        if(isNull(personalRecords))
+            throw new EntityNotFoundException("Resource not available");
         if(!personalRecordsServiceImpl.hasAccess(personalRecords))
             throw new AccessDeniedException("Access denied");
         return mapToDTOs(repository.findThreeLatestByPersonalRecordsId(personalRecordsId));
@@ -65,6 +74,8 @@ public class CustomRecordServiceImpl implements CustomRecordService {
     @Override
     public String deleteCustomRecord(Long customRecordId) {
         CustomRecord customRecord = getCustomRecordByIdFromRepo(customRecordId);
+        if(isNull(customRecord))
+            throw new EntityNotFoundException("Resource not available");
         if(!hasAccess(customRecord))
             throw new AccessDeniedException("Access denied");
         deleteById(customRecordId);
@@ -74,10 +85,11 @@ public class CustomRecordServiceImpl implements CustomRecordService {
     private boolean hasAccess(CustomRecord customRecord) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         UserDetailsImpl userDetailsImpl = (UserDetailsImpl) authentication.getPrincipal();
-        if(customRecord.getPersonalRecords().getUser().getId().equals(userDetailsImpl.getId()))
-            return true;
-        else
-            return false;
+        return customRecord.getPersonalRecords().getUser().getId().equals(userDetailsImpl.getId());
+    }
+
+    private boolean isNull(Object object) {
+        return object == null;
     }
 
     private CustomRecord save(CustomRecord customRecord) {
